Introduce a test_app method
It isn't clear how to test the before and after serving methods, the
documentation talks about directly calling the ``startup`` and
``shutdown`` methods but this makes managing the app nursery very
hard. Hence a new test_app method that must be used in a context
manager (to define when the startup and shutdown methods are called)
has been added.

Note that the existing test client cannot be used in a context manager
for this purpose as it is already used to preserve (extend) the
lifespan of the request context.

This follows a similar change in Quart.